Rename nsclean lib; reorganize background functions; replace NaNs in median image instead of setting to zero
