+++ docs/jwst/clean_flicker_noise/api_ref.rst
@@ -14,10 +14,13 @@ Complete Developer API
 .. automodapi:: jwst.clean_flicker_noise.autoparam
    :no-inheritance-diagram:
 
+.. automodapi:: jwst.clean_flicker_noise.background_level
+   :no-inheritance-diagram:
+
 .. automodapi:: jwst.clean_flicker_noise.clean_flicker_noise
    :no-inheritance-diagram:
 
-.. automodapi:: jwst.clean_flicker_noise.lib
+.. automodapi:: jwst.clean_flicker_noise.nsclean
    :no-inheritance-diagram:
 
 .. automodapi:: jwst.clean_flicker_noise.tso_median_image
