+++ docs/jwst/clean_flicker_noise/reference_files.rst
@@ -1,4 +1,4 @@
 Reference Files
 ===============
-For some optional processing steps, the ``clean_flicker_noise`` step may use the :ref:`flat_reffile`.
+For some optional processing steps, the ``clean_flicker_noise`` step may use the :ref:`flat_reffile`
 or the PASTASOSS reference file.
